Accept an image list so that heterogeneous-sized images can be picked.
Also, return both a list and a stack since user may now select images of
distinct sizes. If they do, the "legacy" stack returns only the first
one to avoid crashes.

This is particularly useful when picking between e.g. a raw image and
its upscaled version, which is very common in ComfyUI.